+++ apollo-api/src/commonMain/kotlin/com/apollographql/apollo3/api/json/BufferedSourceJsonReader.kt
@@ -16,7 +16,7 @@
 package com.apollographql.apollo3.api.json
 
 import com.apollographql.apollo3.api.Throws
-import com.apollographql.apollo3.api.internal.json.JsonScope
+import com.apollographql.apollo3.api.json.internal.JsonScope
 import com.apollographql.apollo3.exception.JsonDataException
 import com.apollographql.apollo3.exception.JsonEncodingException
 import okio.Buffer
