--- apollo-api/src/commonMain/kotlin/com/apollographql/apollo3/api/json/Closeable.kt
@@ -1,8 +1,0 @@
-package com.apollographql.apollo3.api.json
-
-/**
- * A custom [Closeable] interface so that we can use [use]{} in JS and Kotlin native
- */
-expect interface Closeable {
-  fun close()
-}
