+++ apollo-api/src/commonMain/kotlin/com/apollographql/apollo3/api/json/internal/FileUploadAwareJsonWriter.kt
@@ -1,4 +1,4 @@
-package com.apollographql.apollo3.api.internal.json
+package com.apollographql.apollo3.api.json.internal
 
 import com.apollographql.apollo3.api.Upload
 import com.apollographql.apollo3.api.json.JsonNumber
