+++ apollo-api/src/commonMain/kotlin/com/apollographql/apollo3/api/json/internal/MapJsonWriter.kt
@@ -1,5 +1,6 @@
-package com.apollographql.apollo3.api.internal.json
+package com.apollographql.apollo3.api.json.internal
 
+import com.apollographql.apollo3.annotations.ApolloInternal
 import com.apollographql.apollo3.api.Upload
 import com.apollographql.apollo3.api.json.JsonNumber
 import com.apollographql.apollo3.api.json.JsonWriter
@@ -21,6 +22,7 @@ import com.apollographql.apollo3.api.json.JsonWriter
  *
  * To write to a [okio.BufferedSink], see also [BufferedSinkJsonWriter]
  */
+@ApolloInternal
 class MapJsonWriter : JsonWriter {
   sealed class State {
     class List(val list: MutableList<Any?>) : State()
