--- apollo-api/src/commonMain/kotlin/com/apollographql/apollo3/api/json/use.kt
@@ -1,24 +1,0 @@
-package com.apollographql.apollo3.api.json
-
-/**
- * A version of use that works on JS and native
- */
-inline fun <T : Closeable?, R> T.use(block: (T) -> R): R {
-    var exception: Throwable? = null
-    try {
-        return block(this)
-    } catch (e: Throwable) {
-        exception = e
-        throw e
-    } finally {
-        when {
-            this == null -> {}
-            exception == null -> close()
-            else ->
-                try {
-                    close()
-                } catch (closeException: Throwable) {
-                }
-        }
-    }
-}
